Move List into gc_list.h
